+++ R/scan1coef.R
@@ -132,7 +132,7 @@ scan1coef <-
     model <- match.arg(model)
 
     # deal with the dot args
-    dotargs <- list("...")
+    dotargs <- list(...)
     tol <- grab_dots(dotargs, "tol", 1e-12)
     if(!is_pos_number(tol)) stop("tol should be a single positive number")
     if(model=="binary") {